fix(navbar): restore margin beneath navbar
While building out the demo, the navbar margin was set to zero to allow for
multiple navbars (primary, secondary, utility, etc.) to be stacked together.
This broke default Bootstrap behavior and caused margin/padding to be added to
the content. Instead, I've restored the default bottom margin on navbars and
only removed the bottom margin when navbars are sibling elements.